+++ examples/pc/vscp_user/vscp_config_overwrite.h
@@ -106,6 +106,8 @@ extern "C"
 
 #define VSCP_CONFIG_ENABLE_SEGMENT_TIME_CALLOUT VSCP_CONFIG_BASE_DISABLED
 
+#define VSCP_CONFIG_PROTOCOL_EVENT_NOTIFICATION VSCP_CONFIG_BASE_DISABLED
+
 */
 
 /*******************************************************************************