+++ examples/pc/vscp_user/vscp_portable.h
@@ -46,6 +46,7 @@ This module contains VSCP support functionality, which is used by the core.
  * - VSCP_CONFIG_ERROR_CALLOUT
  * - VSCP_CONFIG_ENABLE_SEGMENT_TIME_CALLOUT
  * - VSCP_DEV_DATA_CONFIG_ENABLE_GUID_STORAGE_EXT
+ * - VSCP_CONFIG_PROTOCOL_EVENT_NOTIFICATION
  *
  * @{
  */
@@ -166,9 +167,27 @@ extern BOOL vscp_portable_isBootloaderRequested(void);
 /**
  * This function provides received VSCP events, except the PROTOCOL class.
  *
- * @param[out]  msg Message
+ * @param[in]   msg Message
  */
 extern void vscp_portable_provideEvent(vscp_RxMessage const * const msg);
+
+#if VSCP_CONFIG_BASE_IS_ENABLED( VSCP_CONFIG_PROTOCOL_EVENT_NOTIFICATION )
+
+/**
+ * This function provides received VSCP PROTOCOL class events.
+ *
+ * Attention: Handling events which the core is waiting for can cause bad
+ * behaviour.
+ * 
+ * @param[in]   msg Message
+ * 
+ * @return Event handled or not. If application handles event, the core won't handle it.
+ * @retval FALSE    Event not handled
+ * @retval TRUE     Event handled
+ */
+extern BOOL vscp_portable_provideProtocolEvent(vscp_RxMessage const * const msg);
+
+#endif  /* VSCP_CONFIG_BASE_IS_ENABLED( VSCP_CONFIG_PROTOCOL_EVENT_NOTIFICATION ) */
 
 #if VSCP_CONFIG_BASE_IS_ENABLED( VSCP_CONFIG_ENABLE_SEGMENT_TIME_CALLOUT )