+++ CHANGELOG.md
@@ -25,6 +25,7 @@
         Please increase your max. number of timers to be able to use this feature.
       - Callout function for every received segment controller heartbeat with timestamp implemented.
       - MCU stored GUID is now supported additional to storing the GUID in code flash or in persistent memory.
+      - Application is now able to handle protocol class events, if configured with VSCP_CONFIG_PROTOCOL_EVENT_NOTIFICATION.
 
 ## 0.6.1